refactor(structure): move built-in components to dedicated builtins package
- Create new nornflow.builtins package for standard components
- Move built-in filters (hosts, groups) to nornflow.builtins.filters
- Move DefaultNornFlowProcessor to nornflow.builtins.processors
- Add direct exports from nornflow.builtins.__init__.py for flatter imports
- Update imports across codebase to use new locations
