+++ nornflow/builtins/__init__.py
@@ -1,0 +1,14 @@
+"""
+Built-in components for NornFlow.
+
+This package contains the default filters and processors included with NornFlow.
+"""
+
+from nornflow.builtins.processors import DefaultNornFlowProcessor
+from nornflow.builtins.filters import hosts, groups
+
+__all__ = [
+    "DefaultNornFlowProcessor",
+    "hosts",
+    "groups",
+]
