+++ nornflow/cli/run.py
@@ -160,7 +160,7 @@ def parse_processors(value: str | None) -> list[dict[str, Any]]:
     Parse a string of processor configurations into a list of processor configs.
 
     Format examples:
-      - Single processor: "class='nornflow.processors.DefaultNornFlowProcessor',args={}"
+      - Single processor: "class='nornflow.builtins.processors.DefaultNornFlowProcessor',args={}"
       - Multiple processors:
         "class='module.Processor1',args={};class='module.Processor2',args={'key':'value'}"
