+++ nornflow/cli/samples/nornflow.yaml
@@ -65,5 +65,5 @@ dry_run: False
 #   - Multiple processors: --processors "class='package.Processor1',args={};class='package.Processor2',args={'key':'value'}"
 #
 processors:
-  - class: "nornflow.processors.DefaultNornFlowProcessor"
+  - class: "nornflow.builtins.DefaultNornFlowProcessor"
     args: {}
